+++ Changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * `[tool.maturin]` now supports `sdist-include = ["path/**/*"]` to
 include arbitrary files in source distributions ([#296](https://github.com/PyO3/maturin/pull/296)).
  * cffi is installed if it's missing and python is running inside a virtualenv.
+ * Add support for PyO3 `0.12`'s `PYO3_PYTHON` environment variable. [#331](https://github.com/PyO3/maturin/pull/331)
 
 ## 0.8.0 - 2020-04-03